+++ frontend/src/App.js
@@ -1,6 +1,6 @@
 import './App.css';
 import {
-  BrowserRouter as Router, Route, Routes,
+  BrowserRouter as Router, Route, Routes, Navigate
 }
   from 'react-router-dom';
 import Login from './components/auth/Login';
@@ -12,6 +12,7 @@ function App() {
     <div className="App">
       <Router>
         <Routes>
+          <Route exact path="/" element={<Navigate to="/login" />} />
           <Route path="/login" element={(<Login />)} />
           <Route path="/signup" element={(<Signup />)} />
           <Route path="/inbox" element={(<Inbox />)} />
