+++ frontend/src/components/auth/Login.jsx
@@ -1,13 +1,14 @@
 import React, { useState } from "react";
+import { useNavigate } from "react-router-dom";
 import Container from 'react-bootstrap/Container';
 import Form from "react-bootstrap/Form";
 import Button from "react-bootstrap/Button";
-import "./Signup.css";
+import "./Login.css";
 
 
 // https://sst.dev/chapters/create-a-login-page.html
 
-export default function Signup() {
+export default function Login() {
     " State variables, stores the user inputs"
     const [email, setEmail] = useState("");
     const [username, setUsername] = useState("");
@@ -17,16 +18,21 @@ export default function Signup() {
         return email.length > 0 && password.length > 0;
     }
 
+    const navigate = useNavigate();
+
     function handleSubmit(event) {
         event.preventDefault();
+        navigate("/inbox");
     }
-
+    function handleSignUp(event) {
+        navigate("/signup");
+    }
     return (
         <Container>
 
             <Form className="Login">
                 <img className="Logo" src="" alt=""/>
-                <h1>Sign Up</h1>
+                <h1>Sign In</h1>
                 <Form.Group size="lg" controlId="email">
                 <Form.Label>Email</Form.Label>
                 <Form.Control
@@ -44,16 +50,11 @@ export default function Signup() {
                     onChange={(e) => setPassword(e.target.value)}
                 />
                 </Form.Group>
-                <Form.Group size="lg" controlId="password">
-                <Form.Label>Confirm Password</Form.Label>
-                <Form.Control
-                    type="password"
-                    value={password}
-                    onChange={(e) => setPassword(e.target.value)}
-                />
-                </Form.Group>
-                <Button block="true" size="lg" type="submit" disabled={!validateForm()}>
+                <Button onClick={handleSubmit} block="true" size="lg" type="submit" disabled={!validateForm()}>
                     Submit
+                </Button>
+                <Button onClick={handleSignUp}>
+                    Sign Up
                 </Button>
             </Form>
         </Container>
